Rename Eswatini from Swaziland
