+++ test/countries/test_eswatini.py
@@ -16,7 +16,7 @@ from datetime import date
 import holidays
 
 
-class TestSwaziland(unittest.TestCase):
+class TestEswatini(unittest.TestCase):
     def setUp(self):
         self.holidays = holidays.SZ()
